imap_service/requirements.txt
flask
imapclient
pyzmail36
python-dotenv
requests